requirements.txt
streamlit==1.51.0
opencv-python-headless==4.8.1.78
numpy==1.27.0
Pillow==10.0.0
